requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
python-telegram-bot==13.4.1
requests==2.25.1
